fix(ops): send plain floats over the gradcheck multiprocessing queue
The rank returned a rel dict holding 0-dim tensors (loss accumulated with
tensor .sum(), never .item()). On Linux spawn the queue shards tensor storage
via fd sharing; the worker exits right after put, so the parent hits EOFError
reaching the gone resource_sharer — the ubuntu CI failure. Force the loss to
float and wrap every rel value in float(); the queue now carries no tensors.

+++ tests/gdn_cp_gradcheck_world2.py
@@ -86,7 +86,7 @@ def _global_loss(A, Bb, GT, HT):
     for c in range(NCHUNK):
         total = total + (GT[c] * pa).sum() + (HT[c] * pc).sum()
         pa, pc = A[c] @ pa, A[c] @ pc + Bb[c]
-    return total
+    return float(total)
 
 
 def _decay_pairs(A):
@@ -147,7 +147,8 @@ def _rank(rank: int, decay_a: bool, no_compose: bool) -> dict:
                 f_minus = _global_loss(args_a, args_b, GT, HT)
                 num = (f_plus - f_minus) / (2 * STEP)
                 an = an_t.reshape(-1)[j].item()
-                rel[kind] = max(rel[kind], abs(num - an) / max(abs(num), abs(an), 1e-9))
+                rel[kind] = max(rel[kind],
+                               float(abs(num - an) / max(abs(num), abs(an), 1e-9)))
     dist.barrier()
     return rel
